pyproject.toml
[tool.poetry]
name = "learn-control-backend"
version = "0.1.0"
description = ""
authors = ["pedro-rodrigues18 <hrpereira.pedro@gmail.com>"]
readme = "README.md"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = { extras = ["standard"], version = "^0.30.6" }
pyserial = "^3.5"
pydantic = "^2.8.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
